Example solution to save state
When save is being applied the theme color data is defaulting the WP Core
color palette. This causes a mismatch in what the editor expects and
what is saved in the post content.

In the save state this can be tested by hard coding the values for the
color demo array being passed in.